+++ internal/weather/weather.go
@@ -20,21 +20,21 @@ type CurrentUnits struct {
 type Current struct {
 	Time               string  `json:"time"`
 	Interval           int     `json:"interval"`
-	Temperature2m      float32 `json:"temperature_2m"`
+	Temperature2m      float64 `json:"temperature_2m"`
 	RelativeHumidity2m int     `json:"relative_humidity_2m"`
-	Rain               float32 `json:"rain"`
-	Showers            float32 `json:"showers"`
-	Snowfall           float32 `json:"snowfall"`
+	Rain               float64 `json:"rain"`
+	Showers            float64 `json:"showers"`
+	Snowfall           float64 `json:"snowfall"`
 }
 
 type Weather struct {
-	Latitude             float32      `json:"latitude"`
-	Longitude            float32      `json:"longitude"`
-	GenerationTimeMs     float32      `json:"generation_time_ms"`
+	Latitude             float64      `json:"latitude"`
+	Longitude            float64      `json:"longitude"`
+	GenerationTimeMs     float64      `json:"generation_time_ms"`
 	UTCOffsetSeconds     int          `json:"utc_offset_seconds"`
 	Timezone             string       `json:"timezone"`
 	TimezoneAbbreviation string       `json:"timezone_abbreviation"`
-	Elevation            float32      `json:"elevation"`
+	Elevation            float64      `json:"elevation"`
 	CurrentUnits         CurrentUnits `json:"current_units"`
 	Current              Current      `json:"current"`
 }
@@ -47,7 +47,7 @@ func (w Weather) Validate() (validation.ValidationProblems, error) {
 const basePath = "https://api.open-meteo.com/v1/forecast"
 const fields = "temperature_2m,relative_humidity_2m,rain,showers,snowfall"
 
-func ForLatLon(lat float32, lon float32) (Weather, error) {
+func ForLatLon(lat float64, lon float64) (Weather, error) {
 	weather := Weather{}
 
 	endpoint := fmt.Sprintf("%s?current=%s&latitude=%.2f&longitude=%.2f",
