+++ internal/weather/weather_test.go
@@ -5,8 +5,8 @@ import (
 )
 
 func Test(t *testing.T) {
-	const lat float32 = 8.716667
-	const lon float32 = 167.733333
+	const lat float64 = 8.716667
+	const lon float64 = 167.733333
 
 	t.Run("runs successfully", func(t *testing.T) {
 		if _, err := ForLatLon(lat, lon); err != nil {
